Add: Dictionary to convert numbers to words
Created an initial dictionary of number and string pairs, as a simple way to
convert from one representation to the other.  Only including the minimum of
sample data in case the dictionary is not the right structure.

Added test to cover first two pieces of sample data from the problem
description.