REPL spike: word-sequence function
Built a word-sequence function in the REPL, using the CIDER debugger.  Testing
in the REPL with some sample data.

Added a `:pre` precondition assertion that the number passed to the function
should be greater than or equal to zero.  So any minus numbers passed as an
argument will cause an assertion error.

`Assert failed: (<= 0 number)`