Refactor: number-sequence
Parses the digits in a number and returns individual numbers that include the
number level of their position.

The number is converted to a string, so it can be treated like a sequence.

Each digit is converted to the correct number level by replacing all proceeding
digits with zero.

A sequence of numbers for each digit is returned.