Test: number-sequence - zero intermediate values
Test the number-sequence function for numbers that have one or more zero values
between the start and end digits of the number.

Numbers that start with a zero at number levels are not pronounced, so we should
not include them in the returned sequence.

Changed testing assertions so all assertions testing the same function are
together.