Refactor: dictionaries namespace and lookups
Defined a namespace for the hash-maps that contain the dictionary lookups.

Dictionary lookups added for single-digit, teens and levels of ten numbers.  So
lookups will cover numbers from 0 to 99.

Defined a `british-english-dictionary` by combining the relevant dictionary
hash-maps.

Using a separate namespace as there is quite a lot of declarative code that does
not need to be directly viewable in the main namespace and avoids clutter and
scrolling through the code.