Refactor: number-levels dictionary and word-sequence function
* number-levels collection
changed structure to a hash-map so number level names can be looked up by the
size of the string that represents the positional number.

* word-sequence function
for numbers less than 100 (size 2 or less) then use the dictionary to find the
number.

Numbers 100 or greater will get the first number and post-fix the positional
number level to the end of that number.